Hardcoded Strings shifted to AuthEndPoints
-> overloaded requestAccess to avoid casting and more independent code

->renamed file and better conventions
